5_rest_api/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.11.0
django-environ==0.12.0
psycopg2-binary==2.9.10
uvicorn==0.32.1
faker==19.11.0
ipdb==0.13.13
drf-yasg==1.21.10